requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
django-recaptcha==4.0.0
django-tailwind==3.8.0
django-templated-mail==1.1.1
django-tinymce==3.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
fontawesomefree==6.5.1
idna==3.6
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
razorpay==1.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
urllib3==2.2.1
